requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2025.4
Django==5.0.2
django-appconf==1.1.0
django-ckeditor==6.7.2
django-cloudinary-storage==0.3.0
django-compressor==4.5.1
django-crispy-forms==2.3
django-environ==0.12.0
django-js-asset==3.1.2
django-sitemap-generate==0.7.0
gunicorn==21.2.0
pillow==11.1.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
rcssmin==1.1.2
rjsmin==1.2.2
sqlparse==0.5.3
typing_extensions==4.13.1
tzdata==2025.2
whitenoise==6.6.0
cloudinary==1.36.0
